re-structure gateway devices
