+++ miio/gateway/__init__.py
@@ -1,0 +1,8 @@
+"""Xiaomi Gateway implementation using Miio protecol."""
+
+# flake8: noqa
+from .alarm import Alarm
+from .gateway import Gateway
+from .light import Light
+from .radio import Radio
+from .zigbee import Zigbee
