+++ miio/gateway/devices/__init__.py
@@ -1,0 +1,7 @@
+"""Xiaomi Gateway subdevice base class."""
+
+# flake8: noqa
+from .light import LightBulb
+from .switch import Switch
+
+from .subdevice import SubDevice, SubDeviceInfo  # isort:skip
